App, PetSearch - lift state up to prevent unnecessary re-renders
