+++ web/src/App.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
 import logo from './logo.svg';
 import './App.css';
 import PetSearch from './components/PetSearch';
@@ -7,10 +7,34 @@ import PetList from './components/PetList';
 const App = () => {
   const [pets, setPets] = useState([]);
   const [error, setError] = useState(null);
-  const [filteredPets, setFilteredPets] = useState([]);
+  const [searchTerm, setSearchTerm] = useState("");
+  const [petType, setPetType] = useState("");
 
-  const handleResultsChange = useCallback((nextPets) => {
-    setFilteredPets(nextPets || []);
+  const filteredPets = useMemo(() => {
+    return pets
+      .filter((pet) => {
+        if (!searchTerm) return true;
+        const name = (pet?.name || "").toString();
+        return name.toLowerCase().includes(searchTerm.toLowerCase());
+      })
+      .filter((pet) => {
+        if (!petType) return true;
+        const type = (pet?.type || "").toString();
+        return type.toLowerCase() === petType.toLowerCase();
+      });
+  }, [pets, searchTerm, petType]);
+
+  const handleSearchTermChange = useCallback((e) => {
+    setSearchTerm(e.target.value);
+  }, []);
+
+  const handlePetTypeChange = useCallback((type) => {
+    setPetType(type);
+  }, []);
+
+  const handleClearSearch = useCallback(() => {
+    setSearchTerm("");
+    setPetType("");
   }, []);
 
   useEffect(() => {
@@ -41,7 +65,14 @@ const App = () => {
         <img src={logo} className="App-logo" alt="logo" />
       </header>
       <div className="App-content">
-        <PetSearch data-testid="pet-search" pets={pets} onResultsChange={handleResultsChange} />
+        <PetSearch 
+          data-testid="pet-search" 
+          searchTerm={searchTerm}
+          petType={petType}
+          onSearchTermChange={handleSearchTermChange}
+          onPetTypeChange={handlePetTypeChange}
+          onClearSearch={handleClearSearch}
+        />
         <PetList data-testid="pet-list" error={error} pets={filteredPets} />
       </div>
     </div>
